+++ practice-app/server/src/main/java/com/bounswe/group8/controller/PostController.java
@@ -1,6 +1,6 @@
 package com.bounswe.group8.controller;
 
-import com.bounswe.group8.payload.PostCreateRequest;
+import com.bounswe.group8.payload.request.PostCreateRequest;
 import com.bounswe.group8.payload.dto.PostDto;
 import com.bounswe.group8.service.PostService;
 import lombok.AccessLevel;
